Add search by keyword features
Using Vuxe retrieving data features "where" to query the model with specific field and value.

Field that involve was title , description and type.

Because we are using the direct features that comes with vuex and simply change  and putting extra query on an existing code, The querying functionality required the exact string of respective field value.

FYI, in this commit, prettier do get involve by removing extra commas at the end of an object and also removing () inside a fat arrow function. So you will see alot of code changes